[refactor - loginpage]
