+++ src/components/Login/LoginBtn.js
@@ -4,14 +4,17 @@ import { useGoogleLogin, useGoogleLogout } from 'react-google-login'
 import { GOOGLE_ID } from 'config/config.json'
 import { useStores } from 'stores'
 import { refreshTokenSetup } from 'utils/refreshLoginSetup'
-import './newLogin.scss'
+import 'components/Login/login.scss'
 import googleIcon from 'assets/images/google@2x.png'
 
-const LoginBtn = observer(() => {
-  const { userStore: { accessToken } } = useStores()
+const LoginBtn = observer(({ history }) => {
+  const {
+    userStore: { accessToken },
+  } = useStores()
   const onSuccess = (res) => {
     console.log('login success', res)
     refreshTokenSetup(res)
+    history.push('/signup')
   }
   const onFailure = (err) => {
     console.log('fail', err)
@@ -25,18 +28,18 @@ const LoginBtn = observer(() => {
     onFailure,
     clientId: GOOGLE_ID,
     isSignedIn: true,
-    accessType: 'offline'
+    accessType: 'offline',
   })
 
   const { signOut } = useGoogleLogout({
     clientId: GOOGLE_ID,
     onLogoutSuccess,
-    onFailure
+    onFailure,
   })
 
   return (
     <button onClick={accessToken.length === 0 ? signIn : signOut} className={'loginBtn'}>
-      <img className={'googleIcon'} src={googleIcon} alt={'google'}/>
+      <img className={'googleIcon'} src={googleIcon} alt={'google'} />
       <span className={'loginBtnText'}>{accessToken.length === 0 ? 'LOGIN with Google' : 'LOGOUT'}</span>
     </button>
   )
