--- src/components/Login/newLogin.scss
@@ -1,65 +1,0 @@
-//@import url(https://fonts.googleapis.com/earlyaccess/notosanskr.css);
-//
-//body {
-//  font-family: 'Noto Sans CJK KR', Noto Sans KR, sans-serif;
-//}
-
-.loginSection {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: center;
-  float: right;
-  width: 25%;
-  height: 835px;
-  margin-right: 11%;
-  background-image: linear-gradient(to top, #209cff, #68e0cf);
-  z-index: 1;
-
-  .mainLogo {
-    width: 322px;
-    margin-top: 160px;
-  }
-  .loginBtn {
-    display: flex;
-    flex-flow: row nowrap;
-    justify-content: center;
-    align-items: center;
-    background: #ffffff;
-    width: 322px;
-    height: 44px;
-    margin-top: 120px;
-    padding: 0;
-    border-radius: 10px;
-    border: 0;
-    overflow: hidden;
-    text-align: center;
-    font-weight: bold;
-    cursor: pointer;
-
-    &:focus {
-      outline: none;
-    }
-    .googleIcon {
-      margin: 0;
-      width: 25px;
-    }
-    .loginBtnText {
-      margin-left: 10px;
-      margin-bottom: 1px;
-      font-size: 20px;
-      color: #3f3d3d;
-    }
-  }
-}
-
-@media screen and (min-width: 1440px) and (max-width: 1920px) {
-  .loginSection {
-    //width: 25%;
-    //height: 835px;
-    //margin-right: 11%;
-    .mainLogo {
-      width: 70%;
-    }
-  }
-}
