--- src/pages/loginPage.scss
@@ -1,40 +1,0 @@
-.container {
-    width: 100%;
-    height: 100%;
-    display: block;
-    overflow: hidden;
-    //display: grid;
-    //grid-template-rows: 635px 345px;
-}
-
-.background {
-    width: 100%;
-    height: 835px;
-    background: url('../assets/images/login_back@3x.png') center 0 no-repeat;
-    background-size: contain;
-    margin-top: 60px;
-    z-index: 0;
-}
-
-@media screen and (min-width: 1440px) and (max-width: 1920px) {
-    .background {
-        width: 100%;
-        background-size: auto 835px;
-    }
-}
-
-// .gradientBox{
-//     width:520px;
-//     height:700px;
-//     background-image: linear-gradient(to top, #209cff, #68e0cf);
-//     position: absolute;
-//     z-index: 1;
-// }
-
-// .logo{
-//     background-repeat: no-repeat;
-//     background-image: url('../assets/images/logo.png');
-//     background-position: center;
-//     position: absolute;
-//     z-index: 2;
-// }
